flatten structure
