--- samples/callbacks/Version_3/WebSender/Handlers/ObjectResponseMessageHandler.cs
@@ -1,14 +1,0 @@
-﻿namespace WebSender.Handlers
-{
-    using NServiceBus;
-
-    #region Web_ObjectResponseMessageHandler
-    public class ObjectResponseMessageHandler : IHandleMessages<ObjectResponseMessage>
-    {
-        public void Handle(ObjectResponseMessage message)
-        {
-        }
-    }
-
-    #endregion
-}
