+++ samples/callbacks/Version_3/WebSender/ObjectResponseMessageHandler.cs
@@ -1,0 +1,11 @@
+﻿using NServiceBus;
+
+#region Web_ObjectResponseMessageHandler
+public class ObjectResponseMessageHandler : IHandleMessages<ObjectResponseMessage>
+{
+    public void Handle(ObjectResponseMessage message)
+    {
+    }
+}
+
+#endregion
