+++ samples/callbacks/Version_3/WebSender/WebSender.csproj
@@ -101,13 +101,12 @@
     <None Include="Views\Home\SendIntMessage.cshtml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="App_Start\RouteConfig.cs" />
-    <Compile Include="App_Start\ServiceBus.cs" />
-    <Compile Include="Controllers\HomeController.cs" />
+    <Compile Include="ServiceBus.cs" />
+    <Compile Include="HomeController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Handlers\ObjectResponseMessageHandler.cs" />
+    <Compile Include="ObjectResponseMessageHandler.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Views\web.config" />
