--- samples/callbacks/Version_4/WebSender/App_Start/RouteConfig.cs
@@ -1,19 +1,0 @@
-﻿using System.Web.Mvc;
-using System.Web.Routing;
-
-namespace WebSender
-{
-    public class RouteConfig
-    {
-        public static void RegisterRoutes(RouteCollection routes)
-        {
-            routes.IgnoreRoute("{resource}.axd/{*pathInfo}");
-
-            routes.MapRoute(
-                name: "Default",
-                url: "{controller}/{action}/{id}",
-                defaults: new { controller = "Home", action = "Index", id = UrlParameter.Optional }
-            );
-        }
-    }
-}
