+++ samples/callbacks/Version_4/WebSender/Views/web.config
@@ -16,7 +16,6 @@
         <add namespace="System.Web.Mvc.Ajax" />
         <add namespace="System.Web.Mvc.Html" />
         <add namespace="System.Web.Routing" />
-        <add namespace="WebSender" />
       </namespaces>
     </pages>
   </system.web.webPages.razor>
