+++ samples/callbacks/Version_4/WebSender/WebSender.csproj
@@ -101,9 +101,8 @@
     <None Include="Views\Home\SendIntMessage.cshtml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="App_Start\RouteConfig.cs" />
-    <Compile Include="App_Start\ServiceBus.cs" />
-    <Compile Include="Controllers\HomeController.cs" />
+    <Compile Include="ServiceBus.cs" />
+    <Compile Include="HomeController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
