+++ samples/callbacks/Version_5/WebSender/HomeController.cs
@@ -1,0 +1,49 @@
+﻿using System.Threading.Tasks;
+using System.Web.Mvc;
+
+public class HomeController : Controller
+{
+    public ActionResult Index()
+    {
+        return View();
+    }
+
+    #region Web_SendEnumMessage
+
+    public async Task<ActionResult> SendEnumMessage()
+    {
+        EnumMessage message = new EnumMessage();
+        Task<Status> statusTask = ServiceBus.Bus.Send("Samples.Callbacks.Receiver", message)
+            .Register<Status>();
+
+        return View("SendEnumMessage", await statusTask);
+    }
+
+    #endregion
+
+    #region Web_SendIntMessage
+
+    public async Task<ActionResult> SendIntMessage()
+    {
+        IntMessage message = new IntMessage();
+        Task<int> responseTask = ServiceBus.Bus.Send("Samples.Callbacks.Receiver", message)
+            .Register();
+
+        return View("SendIntMessage", await responseTask);
+    }
+
+    #endregion
+
+    #region Web_SendObjectMessage
+
+    public async Task<ActionResult> SendObjectMessage()
+    {
+        ObjectMessage message = new ObjectMessage();
+        Task<ObjectResponseMessage> responseTask = ServiceBus.Bus.Send("Samples.Callbacks.Receiver", message)
+            .Register(completion => (ObjectResponseMessage) completion.Messages[0]);
+
+        return View("SendObjectMessage", await responseTask);
+    }
+
+    #endregion
+}
