+++ samples/callbacks/Version_5/WebSender/ServiceBus.cs
@@ -1,0 +1,28 @@
+﻿using NServiceBus;
+
+public class ServiceBus
+{
+    public static IBus Bus { get; private set; }
+
+    static readonly object padlock = new object();
+
+    public static void Init()
+    {
+        if (Bus != null)
+            return;
+
+        lock (padlock)
+        {
+            if (Bus != null)
+                return;
+
+            BusConfiguration busConfiguration = new BusConfiguration();
+            busConfiguration.EndpointName("Samples.Callbacks.WebSender");
+            busConfiguration.UseSerialization<JsonSerializer>();
+            busConfiguration.UsePersistence<InMemoryPersistence>();
+            busConfiguration.EnableInstallers();
+
+            Bus = NServiceBus.Bus.Create(busConfiguration).Start();
+        }
+    }
+}
